Moved Edit and Add Forms in to a separate File.
On the one hand, it's much easier to read and unterstand, than one big task.py file and on the other hand
adapting ITask on a task is not possible, probably because of the directives which are used in the task.py.
