+++ opengever/task/task.py
@@ -20,12 +20,11 @@ from opengever.ogds.base.utils import get_current_org_unit
 from opengever.ogds.base.utils import ogds_service
 from opengever.task import _
 from opengever.task import util
-from opengever.task.activities import TaskAddedActivity
 from opengever.task.interfaces import ITaskSettings
 from opengever.task.validators import NoCheckedoutDocsValidator
 from plone import api
 from plone.dexterity.content import Container
-from plone.directives import form, dexterity
+from plone.directives import form
 from plone.indexer.interfaces import IIndexer
 from plone.registry.interfaces import IRegistry
 from Products.CMFCore.permissions import View
@@ -33,7 +32,6 @@ from Products.CMFCore.utils import _mergedLocalRoles
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_unicode
 from z3c.form import validator
-from z3c.form.interfaces import HIDDEN_MODE
 from z3c.relationfield.schema import RelationChoice
 from z3c.relationfield.schema import RelationList
 from zope import schema
@@ -375,57 +373,6 @@ def deadline_default_value(data):
 @form.default_value(field=ITask['responsible_client'])
 def responsible_client_default_value(data):
     return get_current_org_unit().id()
-
-
-# XXX
-# setting the default value of a RelationField does not work as expected
-# or we don't know how to set it.
-# thus we use an add form hack by injecting the values into the request.
-
-class AddForm(dexterity.AddForm):
-    grok.name('opengever.task.task')
-
-    def update(self):
-        # put default value for relatedItems into request
-        paths = self.request.get('paths', [])
-        if paths:
-            self.request.set('form.widgets.relatedItems', paths)
-        # put default value for issuer into request
-        portal_state = getMultiAdapter((self.context, self.request),
-                                       name=u"plone_portal_state")
-        member = portal_state.member()
-        if not self.request.get('form.widgets.issuer', None):
-            self.request.set('form.widgets.issuer', [member.getId()])
-        super(AddForm, self).update()
-
-        # omit the responsible_client field and adjust the field description
-        # of the responsible field if there is only one orgunit configured.
-        if not ogds_service().has_multiple_org_units():
-            self.groups[0].widgets['responsible_client'].mode = HIDDEN_MODE
-            self.groups[0].widgets['responsible'].field.description = _(
-                u"help_responsible_single_client_setup", default=u"")
-
-    def createAndAdd(self, data):
-        task = super(AddForm, self).createAndAdd(data=data)
-        activity = TaskAddedActivity(task, self.request, self.context)
-        activity.record()
-        return task
-
-
-class EditForm(dexterity.EditForm):
-    """Standard EditForm, just require the Edit Task permission"""
-    grok.context(ITask)
-    grok.require('opengever.task.EditTask')
-
-    def update(self):
-        super(EditForm, self).update()
-
-        # omit the responsible_client field and adjust the field description
-        # of the responsible field if there is only one client configured.
-        if not ogds_service().has_multiple_org_units():
-            self.groups[0].widgets['responsible_client'].mode = HIDDEN_MODE
-            self.groups[0].widgets['responsible'].field.description = _(
-                u"help_responsible_single_client_setup", default=u"")
 
 
 def related_document(context):
